ML/requirements.txt
opencv-python>=4.7.0
ultralytics>=8.0.0
pymongo>=4.3.3
boto3>=1.26.0
tqdm>=4.65.0
python-dotenv>=1.0.0
numpy>=1.24.0
matplotlib>=3.7.0 